# lib/api-zod/src/generated/types/activeDay.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ActiveDay {
  date: string;
  aItemCount: number;
  bItemCount: number;
}