# lib/api-zod/src/generated/types/activityEntry.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantSlot } from './participantSlot';

export interface ActivityEntry {
  itemId: string;
  content: string;
  categoryId: string;
  categoryTitle: string;
  date: string;
  slot: ParticipantSlot;
  participantName: string | null;
  createdAt: string;
}