# lib/api-zod/src/generated/types/categoryWithItems.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Item } from './item';
import type { ParticipantSlot } from './participantSlot';

export interface CategoryWithItems {
  id: string;
  title: string;
  date: string;
  slot: ParticipantSlot;
  items: Item[];
}