# lib/api-zod/src/generated/types/claimParticipantBody.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantSlot } from './participantSlot';

export interface ClaimParticipantBody {
  slot: ParticipantSlot;
  /**
     * @minLength 1
     * @maxLength 60
     */
  name: string;
  /** @pattern ^[0-9]{4,8}$ */
  pin: string;
}