# lib/api-zod/src/generated/types/createCategoryBody.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantSlot } from './participantSlot';

export interface CreateCategoryBody {
  slot: ParticipantSlot;
  date: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  title: string;
}