# lib/api-zod/src/generated/types/createItemBody.ts
/**
 * Generated by orval v8.9.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ParticipantSlot } from './participantSlot';

export interface CreateItemBody {
  slot: ParticipantSlot;
  categoryId: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  content: string;
  date: string;
}